requirements.txt
dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-speech
yt-dlp
moviepy
